result edit delete instant authentification

+++ rowsite/row/views.py
@@ -258,6 +258,10 @@ def result_add(request, piece_id=None, athlete_id=None):
 @login_required
 def result_edit(request, id):
     result = get_object_or_404(Result, pk=id)
+    athlete = result.athlete
+    user_athlete = Athlete.objects.get(user=request.user)
+    if not user_coxswain_coach(user_athlete, athlete):
+        return render(request, 'row/denied.html', {})
     if request.method == 'POST':
         form = ResultForm(request.POST)
         if form.is_valid():
@@ -276,6 +280,10 @@ def result_edit(request, id):
 @login_required
 def result_delete(request, id):
     result = get_object_or_404(Result, pk=id)
+    athlete = result.athlete
+    user_athlete = Athlete.objects.get(user=request.user)
+    if not user_coxswain_coach(user_athlete, athlete):
+        return render(request, 'row/denied.html', {})
     result.delete()
     if request.GET and request.GET["next"]:
         return HttpResponseRedirect(request.GET["next"])
